### 2023-06-15 #5558 @lauraboeschoten
Dear editor,
Potential reviewers from the list could be anibalsolon or Sirjla since they
are familiar with elixir.
Furthermore, based on familiarity with elixir and/or some conceptual
knowledge about the idea of data donation, I would suggest the following:
cskaandorp, chartgerink (also editor at JOSS) and NiekPas .
Kind regards,
Laura Boeschoten

Op wo 14 jun 2023 om 13:18 schreef The Open Journals editorial robot <
***@***.***>:

> *Submitting author:* @lauraboeschoten <https://github.com/lauraboeschoten>
> (Laura Boeschoten <http://orcid.org/0000-0002-3536-0474>)
> *Repository:* https://github.com/eyra/port
> *Branch with paper.md* (empty if default branch): d3I-jboss-paper
> *Version:* v1.0.1
> *Editor:* Pending
> *Reviewers:* Pending
> *Managing EiC:* Olivia Guest
> Status
>
> [image: status]
> <https://joss.theoj.org/papers/11922a83c07ff2fc0a2cb70d962984aa>
>
> Status badge code:
>
>
> *Author instructions*
>
> Thanks for submitting your paper to JOSS @lauraboeschoten
> <https://github.com/lauraboeschoten>. *Currently, there isn't a JOSS
> editor assigned* to your paper.
>
> @lauraboeschoten <https://github.com/lauraboeschoten> if you have any
> suggestions for potential reviewers then please mention them here in this
> thread (without tagging them with an @). You can search the list of people
> <https://reviewers.joss.theoj.org/lookup> that have already agreed to
> review and may be suitable for this submission.
>
> *Editor instructions*
>
> The JOSS submission bot @editorialbot <https://github.com/editorialbot>
> is here to help you find and assign reviewers and start the main review. To
> find out what @editorialbot <https://github.com/editorialbot> can do for
> you type:
>
> @editorialbot commands
>
