### 2023-06-15 eyra/port#37 @osorensen
Pre-review comment on JOSS submission
Here are a few suggested changes after my first reading of your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5558). Please update the paper before we start the review.

- On line 43, there seems to be a redundant right parenthesis:

- On line 73, abbreviation OSD2F should be introduced:

- On lines 80-81, there should either be a comma after "controller" or you should change "i.e," to "That is,":

- Line 132, change from "scriptextracts" to "script extracts":
