### 2023-06-20 1021edb
tables can be configured with the port.api the be editable or not
