### 2023-06-20 9b526c8
small textual changes after suggestion of editor
