+++ src/framework/types/elements.ts
@@ -1,45 +1,46 @@
 import { isInstanceOf, isLike } from '../helpers'
-import { } from './commands'
+import {} from './commands'
 import { isPropsUIPage, PropsUIPage } from './pages'
 import { isPropsUIPrompt, PropsUIPrompt } from './prompts'
 
 export type PropsUI =
-  PropsUIText |
-  PropsUIButton |
-  PropsUICheckBox |
-  PropsUIRadioItem |
-  PropsUISpinner |
-  PropsUIProgress |
-  PropsUIHeader |
-  PropsUITable |
-  PropsUISearchBar |
-  PropsUIPage |
-  PropsUIPrompt
+  | PropsUIText
+  | PropsUIButton
+  | PropsUICheckBox
+  | PropsUIRadioItem
+  | PropsUISpinner
+  | PropsUIProgress
+  | PropsUIHeader
+  | PropsUITable
+  | PropsUISearchBar
+  | PropsUIPage
+  | PropsUIPrompt
 
 export type PropsUIText =
-  PropsUITextTitle0 |
-  PropsUITextTitle1 |
-  PropsUITextTitle2 |
-  PropsUITextTitle3 |
-  PropsUITextTitle6 |
-  PropsUITextBodyLarge |
-  PropsUITextLabel
+  | PropsUITextTitle0
+  | PropsUITextTitle1
+  | PropsUITextTitle2
+  | PropsUITextTitle3
+  | PropsUITextTitle6
+  | PropsUITextBodyLarge
+  | PropsUITextLabel
 
 export type PropsUIButton =
-  PropsUIButtonPrimary |
-  PropsUIButtonSecundary |
-  PropsUIButtonBack |
-  PropsUIButtonForward |
-  PropsUIButtonIconBack |
-  PropsUIButtonIconForward |
-  PropsUIButtonIcon |
-  PropsUIButtonLabel |
-  PropsUIButtonIconLabel
+  | PropsUIButtonPrimary
+  | PropsUIButtonSecundary
+  | PropsUIButtonBack
+  | PropsUIButtonForward
+  | PropsUIButtonIconBack
+  | PropsUIButtonIconForward
+  | PropsUIButtonIcon
+  | PropsUIButtonLabel
+  | PropsUIButtonIconLabel
 
 // UI
 
-export function isPropsUI (arg: any): arg is PropsUI {
-  return isPropsUIText(arg) ||
+export function isPropsUI(arg: any): arg is PropsUI {
+  return (
+    isPropsUIText(arg) ||
     isPropsUIButton(arg) ||
     isPropsUISpinner(arg) ||
     isPropsUIProgress(arg) ||
@@ -47,12 +48,14 @@ export function isPropsUI (arg: any): arg is PropsUI {
     isPropsUITable(arg) ||
     isPropsUIPage(arg) ||
     isPropsUIPrompt(arg)
+  )
 }
 
 // TEXTS
 
-export function isPropsUIText (arg: any): arg is PropsUIText {
-  return isPropsUITextTitle0(arg) ||
+export function isPropsUIText(arg: any): arg is PropsUIText {
+  return (
+    isPropsUITextTitle0(arg) ||
     isPropsUITextTitle0(arg) ||
     isPropsUITextTitle1(arg) ||
     isPropsUITextTitle2(arg) ||
@@ -63,6 +66,7 @@ export function isPropsUIText (arg: any): arg is PropsUIText {
     isPropsUITextBodyMedium(arg) ||
     isPropsUITextLabel(arg) ||
     isPropsUITextCaption(arg)
+  )
 }
 
 export interface PropsUITextLabel {
@@ -71,7 +75,7 @@ export interface PropsUITextLabel {
   color?: string
   margin?: string
[…]
