+++ tailwind.config.js
@@ -2,9 +2,7 @@
 const plugin = require('tailwindcss/plugin')
 
 module.exports = {
-  content: [
-    './src/**/*.{js,jsx,ts,tsx}'
-  ],
+  content: ['./src/**/*.{js,jsx,ts,tsx}'],
 
   darkMode: 'media',
   theme: {
@@ -194,6 +192,7 @@ module.exports = {
           }
         }
       }
+
       addUtilities(newUtilities)
     })
   ]
