+++ src/framework/types/visualizations.ts
@@ -3,7 +3,6 @@ export type AggregationFunction = 'count' | 'mean' | 'sum' | 'count_pct'
 export interface Axis {
   label?: string
   column: string
-  formatDate?: 'auto' | 'year' | 'year_month' | 'month' | 'day' | 'weekday' | 'day_hour' | 'hour'
 }
 
 export interface YAxis extends Axis {
@@ -12,33 +11,39 @@ export interface YAxis extends Axis {
   secondAxis?: boolean
   z?: string
   zAggregate?: AggregationFunction
+  addZeroes?: boolean
 }
 
-export interface VisualizationLineChart {
+export interface VisualizationInput {
+  x: Axis
+  ys: YAxis[]
+  dateFormat?: DateFormat
+}
+
+export interface VisualizationLineChart extends VisualizationInput {
   type: 'line'
-  x: Axis
-  ys: YAxis[]
 }
 
-export interface VisualizationBarChart {
+export interface VisualizationBarChart extends VisualizationInput {
   type: 'bar'
-  x: Axis
-  ys: YAxis[]
 }
 
 export type AggregateRowIds = Record<string, string[]>
+
 export interface AxisSettings {
   label: string
   secondAxis?: boolean
 }
 
+export type AggregatedData = Record<string, string | AggregateRowIds>[]
+
 export interface VisualizationData {
   type: 'line' | 'bar'
-  tableId: string
-  position: 'top' | 'bottom' | 'table'
-  data: Record<string, string | AggregateRowIds>[]
+  data: AggregatedData
   xKey: AxisSettings
   yKeys: Record<string, AxisSettings>
 }
 
 export type VisualizationType = VisualizationLineChart | VisualizationBarChart
+
+export type DateFormat = 'auto' | 'year' | 'month' | 'day' | 'hour' | 'month_cycle' | 'weekday_cycle' | 'day_cycle' | 'hour_cycle'
