### 2023-06-22 114158d
some examples of line, bar and area graphs
