+++ src/framework/types/visualizations.ts
@@ -1,4 +1,4 @@
-export type AggregationFunction = 'count' | 'mean' | 'sum' | 'count_pct'
+export type AggregationFunction = 'count' | 'mean' | 'sum' | 'count_pct' | 'pct'
 
 export interface Axis {
   label?: string
@@ -14,33 +14,36 @@ export interface YAxis extends Axis {
   addZeroes?: boolean
 }
 
-export interface VisualizationInput {
+export interface ChartProps {
   x: Axis
   ys: YAxis[]
   dateFormat?: DateFormat
   height?: number
 }
 
-export interface VisualizationLineChart extends VisualizationInput {
+export interface LineChartProps extends ChartProps {
   type: 'line'
 }
 
-export interface VisualizationBarChart extends VisualizationInput {
+export interface BarChartProps extends ChartProps {
   type: 'bar'
 }
 
-export interface VisualizationAreaChart extends VisualizationInput {
+export interface AreaChartProps extends ChartProps {
   type: 'area'
 }
 
 export type AggregateRowIds = Record<string, string[]>
 
+export type AggregatedData = Record<string, string | AggregateRowIds>[]
+
 export interface AxisSettings {
   label: string
   secondAxis?: boolean
+  tickerFormat?: TickerFormat
 }
 
-export type AggregatedData = Record<string, string | AggregateRowIds>[]
+export type TickerFormat = 'percent' | 'default'
 
 export interface VisualizationData {
   type: 'line' | 'bar' | 'area'
@@ -49,6 +52,6 @@ export interface VisualizationData {
   yKeys: Record<string, AxisSettings>
 }
 
-export type VisualizationType = VisualizationLineChart | VisualizationBarChart | VisualizationAreaChart
+export type VisualizationType = LineChartProps | BarChartProps | AreaChartProps
 
 export type DateFormat = 'auto' | 'year' | 'quarter' | 'month' | 'day' | 'hour' | 'month_cycle' | 'weekday_cycle' | 'day_cycle' | 'hour_cycle'
