### 2023-06-30 e21233d
script and logic for a version number mechanism
