### 2023-07-10 359bae8
search now also affects figures, and heavy lifting moved to webworker
