+++ src/framework/visualisation/react/ui/workers/visualizationDataWorker.ts
@@ -1,0 +1,176 @@
+import { PropsUITable, TableContext, TableWithContext } from '../../../../types/elements'
+import { VisualizationData, VisualizationType, DateFormat, TickerFormat } from '../../../../types/visualizations'
+
+interface Input {
+  table: TableWithContext
+  visualization: VisualizationType
+}
+
+self.onmessage = (e: MessageEvent<Input>) => {
+  createVisualizationData(e.data.table, e.data.visualization)
+    .then((visualizationData) => {
+      self.postMessage({ status: 'success', visualizationData })
+    })
+    .catch((error) => {
+      console.error(error)
+      self.postMessage({ status: 'error', visualizationData: undefined })
+    })
+}
+
+async function createVisualizationData(table: PropsUITable & TableContext, visualization: VisualizationType): Promise<VisualizationData> {
+  const visualizationData: VisualizationData = {
+    type: visualization.type,
+    xKey: { label: visualization.x.label || visualization.x.column },
+    yKeys: {},
+    data: []
+  }
+
+  // First get the unique values of the x column
+  const rowIds = table.body.rows.map((row) => row.id)
+  let x = getTableColumn(table, visualization.x.column)
+  if (!x || x.length === 0) throw new Error(`X column ${table.id}.${visualization.x.column} not found`)
+  let xSortable: (string | number)[] | null = null // separate variable allows using epoch time for sorting dates
+
+  // ADD CODE TO TRANSFORM TO DATE, BUT THEN ALSO KEEP AN INDEX BASED ON THE DATE ORDER
+  if (visualization.dateFormat) {
+    ;[x, xSortable] = formatDate(x, visualization.dateFormat)
+  }
+
+  const aggregate: Record<string, any> = {}
+  for (let y of visualization.ys) {
+    const aggFun = y.aggregate || 'count'
+    let tickerFormat: TickerFormat = 'default'
+    if (aggFun === 'pct' || aggFun === 'count_pct') tickerFormat = 'percent'
+
+    const yValues = getTableColumn(table, y.column)
+    if (!yValues) throw new Error(`Y column ${table.id}.${y.column} not found`)
+
+    // If group_by column is specified, the columns in the aggregated data will be the unique group_by columns
+    const yGroup = y.group_by ? getTableColumn(table, y.group_by) : null
+
+    // if missing values should be treated as zero, we need to add the missing values after knowing all groups
+    const addZeroes = !!y.addZeroes
+    const groupSummary: Record<string, any> = {}
+    const uniqueGroups = new Set<string>([])
+
+    for (let i = 0; i < x.length; i++) {
+      const xValue = x[i]
+      const yValue = yValues[i]
+      const group = yGroup ? yGroup[i] : y.label || y.column
+      if (addZeroes) uniqueGroups.add(group)
+      const sortBy = xSortable ? xSortable[i] : x[i]
+
+      // calculate group summary statistics. This is used for the mean, pct and count_pct aggregations
+      if (!groupSummary[group]) groupSummary[group] = { n: 0, sum: 0 }
+      if (aggFun === 'count_pct' || aggFun === 'mean') groupSummary[group].n += 1
+      if (aggFun === 'pct') groupSummary[group].sum += Number(yValue) || 0
+
+      // add the AxisSettings for the yKeys in this loop, because we need to get the unique group values from the data (if group_by is used)
[…]
