### 2023-07-11 35eaeea
terminate workers on unmount
