### 2023-07-13 593e1b9
refactoring stuff and moving the data processing outside of the Table component
