### 2023-07-24 #5596 @osorensen
Good to hear that you've started, @leonardojaneis. You can find the full guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

You can also see the most important points at the top of this issue. In particular, I suggest starting by generating your review checklist by writing the following line in a comment in this issue:

`@editorialbot generate my checklist`

You'realso welcome to reach out to me if you have further questions.
