### 2023-08-01 #5596 @kaustavbhattacharjee
@osorensen I have completed my review. Here are the comments for the checkboxes I did not tick:

**Contribution and authorship:** I cannot see any commits from the submitting author ([@lauraboeschoten](https://github.com/lauraboeschoten)) as seen on this page: [https://github.com/eyra/port/graphs/contributors](https://github.com/eyra/port/graphs/contributors)

**Example Usage:** Though there are instructions on how to host it online, there isn't any example/demo website that I can try. It would be great if the authors could add a video of implementing Port on a live website or something similar.

**Community guidelines:** I did not observe any directions on contributing to the software. The authors can mention if they would like new collaborators to open issues on Github with specific labels or contact them directly.

**State of the field:** I could find any comparison with other similar tools (in the paper).

**References:** Some of the cited sources are websites. I think it's a good idea to add the date it was accessed.

**Meta comment:** I would recommend the authors change the title of the React app. It still shows the default "React App" title when viewed in a browser.
