### 2023-08-04 #5596 @leonardojaneis
Hi @osorensen, here is my review.

<hr>

**Contribution and authorship:** the submitting author ([@lauraboeschoten](https://github.com/lauraboeschoten)) ​does not show any commits.

**Community guidelines:** There are no clear guidelines.

**Example usage:** There could be at least one example of use available.

**Summary:** The text offers a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience.

**A statement of need:** the paper does have the section that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work.

**State of the field:** the authors do not describe any comparison to other commonly-used packages. This step is very important because it provides valuable context, insights, and considerations for users, researchers, and developers when evaluating and adopting new software tools.

**Quality of writing:** in general the writing quality is good, I would suggest to include a concluding section summarizing the key points and emphasizing the importance of Port as a software tool for digital data donation. Some sentences may require rephrasing to improve clarity and avoid potential confusion in the Summary section.

**References:** It seems that all the relevant sources have been cited properly in the text. The reference list looks well-organized and complete. Each entry provides the essential details required to locate the source. Reviewer @kaustavbhattacharjee point can be adopted.
