### 2023-08-17 eyra/port#44 @lauraboeschoten
Include example usage
**Comments by reviewers:**

- Example Usage: Though there are instructions on how to host it online, there isn't any example/demo website that I can try. It would be great if the authors could add a video of implementing Port on a live website or something similar.
- Example usage: There could be at least one example of use available.

**I suggest to include in the main text:**

under the tab “Port + Next on the Azure Cloud (Recommended)”: Terraform configuration files to host Port on Azure can be found in the following repository: https://github.com/d3i-infra/infrastructure-d3i-pilot

under the tab “Port API examples” include an additional header with: Example applications of Port:
o	https://github.com/d3i-infra/port-d3i-pilot
o	https://github.com/d3i-infra/port-vu-pilot
o	https://github.com/d3i-infra/port-news-dynamics
o	https://github.com/d3i-infra/port-pilot-netflix
