### 2023-08-17 eyra/port#45 @lauraboeschoten
Change title of React app 
Comment from reviewer:

-	Meta comment: I would recommend the authors change the title of the React app. It still shows the default "React App" title when viewed in a browser.
