### 2023-08-17 eyra/port#46 @lauraboeschoten
Create community guidelines 
The reviewers suggested to create community guidelines:

- Community guidelines: I did not observe any directions on contributing to the software. The authors can mention if they would like new collaborators to open issues on Github with specific labels or contact them directly.
- Community guidelines: There are no clear guidelines.
