+++ D3I_JOSS_paper/paper.bib
@@ -139,7 +139,9 @@
   publisher    = {Zenodo},
   version      = {0.23.0},
   doi          = {10.5281/zenodo.5156931},
-  url          = {https://doi.org/10.5281/zenodo.5156931}
+  url          = {https://doi.org/10.5281/zenodo.5156931},
+   publisher    =	{(Date accessed: 11.08.2023)},
+  urldate = {11/08/23}
 }
 
 @misc{WebAssembly, 
@@ -147,7 +149,9 @@
   author      = {WebAssembly},
   year        = {2021},
   journal     = {WebAssembly},
-  url         = {https://webassembly.org/}
+  url         = {https://webassembly.org/},
+   publisher    =	{(Date accessed: 11.08.2023)},
+  urldate = {11/08/23}
 }
 
 
@@ -156,8 +160,49 @@
   author      = {React},
   year        = {2022},
   journal     = {React},
-  url         = {https://react.dev/}
+  url         = {https://react.dev/},
+   publisher    =	{(Date accessed: 11.08.2023)},
+  urldate = {11/08/23}
 }
 
+@article{Zannettou_2023,
+   abstract = {TikTok is a relatively novel and widely popular media platform. In response to its expanding user base and cultural impact, researchers are turning to study the platform; however, TikTok, like many social media platforms, restricts external access to data. Prior works have acquired data from scraping the platform, user self-reports, and from accounts created by researchers for the study's purpose. Existing techniques, while yielding important insights, contain limitations for gathering large-scale quantitative insights on how real TikTok users behave on the platform. We bridge this research gap by implementing a data donation system to collect TikTok data. Our system leverages users' right to access their data enabled by the EU's GDPR regulation. We recruit 347 TikTok users, ask them to request their data from TikTok, and then use our system to customize, anonymize, and donate their data. We collect 4.9M videos viewed 9.2M times by our participants -- and associated engagement metrics -- to analyze how people consume content on TikTok, how prevalent liking behavior is on TikTok, and whether there are substantial differences across our participants' demographics. We conclude our work by discussing the lessons learned and future avenues for implementing data donation systems, which we believe offer a promising avenue for collecting user behavioral traces to understand social phenomena through the lens of the Web.},
+   author = {Savvas Zannettou and Olivia-Nemes Nemeth and Oshrat Ayalon and Angelica Goetzen and Krishna P. Gummadi and Elissa M. Redmiles and Franziska Roesner},
+   month = {1},
+   title = {Leveraging Rights of Data Subjects for Social Media Analysis: Studying TikTok via Data Donations},
+   url = {http://arxiv.org/abs/2301.04945},
+   year = {2023},
+}
 
+@computer{Pfiffner_2022,
+   author = {Nico Pfiffner and Pim Witlox and Thomas N. Friemel},
+   title = {Data Donation Module},
+   url = {https://github.com/uzh/ddm},
+   urldate = {11/08/23},
+   publisher    =	{(Date accessed: 11.08.2023)},
+   year = {2022}
+}
+
+@inproceedings{GomezOrtega_2021,
[…]
