### 2023-08-21 #5596 @lauraboeschoten
Dear editor and reviewers,
thanks a lot for your feedback! Below you can find our point-by-point response. Some comments only have a response written below, others have changes in the text of the paper. A new version of the paper is now [online](https://github.com/eyra/port/tree/v1.0.1-d3I-jboss-paper/D3I_JOSS_paper). Other points require changes in either the code of Port or on the readme. For these, I made issues. However, as some of our programmers are still on holiday, these are not resolved yet and will be picked up in the coming weeks. You can follow the issues, but I will also send notifications in this thread once issues have been resolved.
I hope this is sufficient and clear enough for now, but do let me know when you have any questions or comments!
Best,
Laura

By @kaustavbhattacharjee and @leonardojaneis:

- No commits from submitting author

Good to point this out. I can provide some background information regarding the complete list of co-authors on this manuscript. As you can infer from the software, it is a tool for data collection. As it is very generic in its set-up, the group consists of applied researchers, methodologists, legal experts and research engineers and software developers employed universities or at private companies. As a methodologist, I am not involved as a programmer in this project. However, I developed the generic framework which Port facilitates and was highly involved in the proof-of-concept of Port, which was developed prior to the version currently under review (see the references). I am currently involved as a project lead, which means that I frequently have design and refinement meetings with the developers, and take responsibility for circling any updates back to the team of researchers and involve them when relevant.

- 	State of the field: I could find any comparison with other similar tools (in the paper).
- 	State of the field: the authors do not describe any comparison to other commonly-used packages. This step is very important because it provides valuable context, insights, and considerations for users, researchers, and developers when evaluating and adopting new software tools.

Thanks for pointing this out, I included the following paragraph in the ‘statement of need’ section: “_In recent years, researchers have implemented the local processing step in multiple ways. Studies by van Driel et al. (2022) and (Kmetty & Németh, 2022) used an approach where DDPs were donated directly, so without a local processing step taking place. However, extensive de-identification procedures were in place to guarantee participant’s privacy (see e.g. Boeschoten et al., 2021). In addition, multiple apps have been developed to enable this local processing step. For example, the app OSD2F (Araujo et al., 2022) allows to select .json files within the DDP and performs certain de-identification steps prior to donation. Alternatively, the app Port (Boeschoten, Mendrik, et al., 2022) allows for more rigorous data processing, as researchers can write a custom Python script that locally performs any requested processing task. While this approach undoubtedly offers greater flexibility, it concurrently places a larger level of responsibility on the researcher. Other apps have been documented in academic literature. However, the feasibility of their reuse for researchers outside their respective institutions or for other platforms as intended is more ambiguous. Examples of such apps are Designerly Data Donation (Gómez Ortega et al., 2021), the Data Donation Module (Pfiffner et al., 2022), the Social Media Donator (Zannettou et al., 2023) and WhatsR (Kohne, 2023)._”

- 	References: Some of the cited sources are websites. I think it's a good idea to add the date it was accessed.
- 	References: It seems that all the relevant sources have been cited properly in the text. The reference list looks well-organized and complete. Each entry provides the essential details required to locate the source. Reviewer @kaustavbhattacharjee point can be adopted.

I included the date visited for the software references.

By @kaustavbhattacharjee:

- Quality of writing: in general the writing quality is good, I would suggest to include a concluding section summarizing the key points and emphasizing the importance of Port as a software tool for digital data donation. Some sentences may require rephrasing to improve clarity and avoid potential confusion in the Summary section.

I included a “Conclusion” paragraph at the end of the manuscript:
“_To summarize, by utilizing GDPR's right of data access, data donation is a promising new approach to collect digital trace data for research purposes. The data donation workflow as introduced by Boeschoten et al., (2022) introduces the idea of locally processing the obtained digital traces at the device of a participant as such that only the digital traces that are of legitimate interest to the researcher are shared. The software introduced in this paper, Port, allows a research to configure a custom data donation study. The research can decide: Which digital platform to investigate, which digital traces to collect, how to present the digital traces to the participant, and what to communicate to the participant throughout this process. These functionalities make Port a generic and useful tool for any researcher interested in collecting digital traces for research purposes._”

By @kaustavbhattacharjee and @leonardojaneis:

-	Example Usage: Though there are instructions on how to host it online, there isn't any example/demo website that I can try. It would be great if the authors could add a video of implementing Port on a live website or something similar.
-	Example usage: There could be at least one example of use available.

I discussed these comments with the development team, but we were quite confused specifically the first comment seems to refer to different  aspects of the tool.

Regarding hosting of the tool. As data donation studies using Port potentially involve special category personal data, high security measures should be in place when used for actual data collection, and relies on infrastructure available to the researcher. Terraform configuration files to host Port on Azure can be found in the following repository: https://github.com/d3i-infra/infrastructure-d3i-pilot. However, as configuration highly depends on the infrastructure available to the researcher, we do not think providing instruction videos on how to do this would be very useful. We can include a link to this repository on Github and keep a list if other configurations come available. We are currently also in the process of integrating into a software foundation which will provide managed hosting solutions as well as manuals for self-hosting.

Regarding examples of use. We have a couple, see:
o	https://github.com/d3i-infra/port-d3i-pilot
o	https://github.com/d3i-infra/port-vu-pilot
o	https://github.com/d3i-infra/port-news-dynamics
o	https://github.com/d3i-infra/port-pilot-netflix

Also note that we host the website https://datadonation.eu/ where we plan to provide more context on how to perform data donation studies in general and refer to external sources such as Python scrips of data donation applications and configuration files.

I added an issue to Github to include these in the readme, please follow: https://github.com/eyra/port/issues/44

By @kaustavbhattacharjee:
- 	Meta comment: I would recommend the authors change the title of the React app. It still shows the default "React App" title when viewed in a browser.

I added an issue to GitHub to change this, please follow:  https://github.com/eyra/port/issues/45

By @kaustavbhattacharjee and @leonardojaneis:
-	Community guidelines: I did not observe any directions on contributing to the software. The authors can mention if they would like new collaborators to open issues on Github with specific labels or contact them directly.
-	Community guidelines: There are no clear guidelines.
We are working on developing community guidelines, I created an issue to Github to add this, please follow: https://github.com/eyra/port/issues/46
