### 2023-08-22 #5596 @osorensen
Thanks for your response, @lauraboeschoten.

Regarding authorship, I've discussed this with other JOSS editors, and we agree that your response to this issue is satisfactory, and in line with our [guidelines for authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship).

Regarding the other issues, I'll await the reviewers' response to them.
