### 2023-09-01 7edc934
changed the title of the react app, added favicon icon
