+++ src/framework/types/elements.ts
@@ -39,7 +39,7 @@ export type PropsUIButton =
 
 // UI
 
-export function isPropsUI(arg: any): arg is PropsUI {
+export function isPropsUI (arg: any): arg is PropsUI {
   return (
     isPropsUIText(arg) ||
     isPropsUIButton(arg) ||
@@ -54,7 +54,7 @@ export function isPropsUI(arg: any): arg is PropsUI {
 
 // TEXTS
 
-export function isPropsUIText(arg: any): arg is PropsUIText {
+export function isPropsUIText (arg: any): arg is PropsUIText {
   return (
     isPropsUITextTitle0(arg) ||
     isPropsUITextTitle0(arg) ||
@@ -76,7 +76,7 @@ export interface PropsUITextLabel {
   color?: string
   margin?: string
 }
-export function isPropsUITextLabel(arg: any): arg is PropsUITextLabel {
+export function isPropsUITextLabel (arg: any): arg is PropsUITextLabel {
   return isInstanceOf<PropsUITextLabel>(arg, 'PropsUITextLabel', ['text', 'color', 'margin'])
 }
 
@@ -86,7 +86,7 @@ export interface PropsUITextCaption {
   color?: string
   margin?: string
 }
-export function isPropsUITextCaption(arg: any): arg is PropsUITextCaption {
+export function isPropsUITextCaption (arg: any): arg is PropsUITextCaption {
   return isInstanceOf<PropsUITextCaption>(arg, 'PropsUITextCaption', ['text', 'color', 'margin'])
 }
 
@@ -96,7 +96,7 @@ export interface PropsUITextBodyLarge {
   color?: string
   margin?: string
 }
-export function isPropsUITextBodyLarge(arg: any): arg is PropsUITextBodyLarge {
+export function isPropsUITextBodyLarge (arg: any): arg is PropsUITextBodyLarge {
   return isInstanceOf<PropsUITextBodyLarge>(arg, 'PropsUITextBodyLarge', [
     'text',
     'color',
@@ -110,7 +110,7 @@ export interface PropsUITextBodyMedium {
   color?: string
   margin?: string
 }
-export function isPropsUITextBodyMedium(arg: any): arg is PropsUITextBodyMedium {
+export function isPropsUITextBodyMedium (arg: any): arg is PropsUITextBodyMedium {
   return isInstanceOf<PropsUITextBodyMedium>(arg, 'PropsUITextBodyMedium', [
     'text',
     'color',
@@ -124,7 +124,7 @@ export interface PropsUITextBodySmall {
   color?: string
   margin?: string
 }
-export function isPropsUITextBodySmall(arg: any): arg is PropsUITextBodySmall {
+export function isPropsUITextBodySmall (arg: any): arg is PropsUITextBodySmall {
   return isInstanceOf<PropsUITextBodySmall>(arg, 'PropsUITextBodySmall', [
     'text',
     'color',
@@ -138,7 +138,7 @@ export interface PropsUITextTitle0 {
   color?: string
   margin?: string
 }
-export function isPropsUITextTitle0(arg: any): arg is PropsUITextTitle0 {
+export function isPropsUITextTitle0 (arg: any): arg is PropsUITextTitle0 {
   return isInstanceOf<PropsUITextTitle0>(arg, 'PropsUITextTitle0', ['text', 'color', 'margin'])
 }
 
@@ -148,7 +148,7 @@ export interface PropsUITextTitle1 {
   color?: string
   margin?: string
 }
-export function isPropsUITextTitle1(arg: any): arg is PropsUITextTitle1 {
+export function isPropsUITextTitle1 (arg: any): arg is PropsUITextTitle1 {
[…]
