+++ src/framework/types/elements.ts
@@ -386,11 +386,10 @@ export interface PropsUITable {
   id: string
   head: PropsUITableHead
   body: PropsUITableBody
-  editable: boolean
   pageSize?: number
 }
 export function isPropsUITable (arg: any): arg is PropsUITable {
-  return isInstanceOf<PropsUITable>(arg, 'PropsUITable', ['editable', 'pageSize', 'id', 'head', 'body'])
+  return isInstanceOf<PropsUITable>(arg, 'PropsUITable', ['pageSize', 'id', 'head', 'body'])
 }
 
 export interface PropsUITableHead {
