+++ src/framework/processing/py/port/script.py
@@ -1,321 +1,65 @@
-import logging
-import json
-import io
-
-import pandas as pd
-
 import port.api.props as props
-import port.helpers as helpers
-import port.unzipddp as unzipddp
-import port.netflix as netflix
-
-
 from port.api.commands import (CommandSystemDonate, CommandUIRender)
 
-LOG_STREAM = io.StringIO()
-
-logging.basicConfig(
-    stream=LOG_STREAM,
-    level=logging.INFO,
-    format="%(asctime)s --- %(name)s --- %(levelname)s --- %(message)s",
-    datefmt="%Y-%m-%dT%H:%M:%S%z",
-)
-
-LOGGER = logging.getLogger("script")
-
-TABLE_TITLES = {
-    "netflix_ratings": props.Translatable(
-        {
-            "en": "Ratings you gave according to Netlix:",
-            "nl": "Jouw beoordelingen volgens Netflix:",
-        }
-    ),
-}
-
-def render_questionnaire(platform, progress):
-    header = props.PropsUIHeader(props.Translatable({"en": "ASD", "nl": "ASD"}))
-    question1 = props.Translatable({"en": "Question?", "nl": "Vraag?"})
-    choices1 = [
-        props.Translatable({"en": "Banana", "nl": "Banaan"}),
-        props.Translatable({"en": "Pear", "nl": "Appel"})
-    ]
-
-    question2 = props.Translatable({"en": "Checkbox Question?", "nl": "Tjekbox Vraag?"})
-    choices2 = [
-        props.Translatable({"en": "Banana", "nl": "Banaan"}),
-        props.Translatable({"en": "Pinaple", "nl": "Ananas"}),
-        props.Translatable({"en": "Pear", "nl": "Appel"})
-    ]
-
-    question3 = props.Translatable({"en": "Open question?", "nl": "Open vraag?"})
-
-    questions = [
-        props.PropsUIQuestionMultipleChoice(question=question1, id=1, choices=choices1),
-        props.PropsUIQuestionMultipleChoiceCheckbox(question=question2, id=2, choices=choices2),
-        props.PropsUIQuestionOpen(question=question3, id=3)
-    ]
-    description = props.Translatable({"en": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.", "nl": "asd"})
-    body = props.PropsUIPromptQuestionnaire(questions=questions, description=description)
-    footer = props.PropsUIFooter(progress)
-
-    page = props.PropsUIPageDonation("ASD", header, body, footer)
-    return CommandUIRender(page)
+import pandas as pd
+import zipfile
 
 
-def process(session_id):
-    LOGGER.info("Starting the donation flow")
-    yield donate_logs(f"{session_id}-tracking")
+def process(sessionId):
+    yield donate(f"{sessionId}-tracking", '[{ "message": "user entered script" }]')
+
+    platforms = ["Twitter", "Facebook", "Instagram", "Youtube"]
+
+    subflows = len(platforms)
+    steps = 2
+    step_percentage = (100/subflows)/steps
 
     # progress in %
-    subflows = 1
-    steps = 2
-    step_percentage = (100 / subflows) / steps
     progress = 0
-    progress += step_percentage
 
-    platform_name = "Netflix"
-    table_list = None
[…]
