### 2023-09-14 eyra/port#46 @AdrienneMendrik
CONTRIBUTING.md added and a contributing section added to the README.
