### 2023-09-15 5f6142a
tried to make all code lint, did not succeed in linting everything, will ask for a review
