+++ src/framework/processing/py/port/script.py
@@ -1,65 +1,289 @@
-import port.api.props as props
-from port.api.commands import (CommandSystemDonate, CommandUIRender)
+import logging
+import json
+import io
 
 import pandas as pd
-import zipfile
+
+import port.api.props as props
+import port.helpers as helpers
+import port.unzipddp as unzipddp
+import port.netflix as netflix
 
 
-def process(sessionId):
-    yield donate(f"{sessionId}-tracking", '[{ "message": "user entered script" }]')
+from port.api.commands import (CommandSystemDonate, CommandUIRender)
 
-    platforms = ["Twitter", "Facebook", "Instagram", "Youtube"]
+LOG_STREAM = io.StringIO()
 
-    subflows = len(platforms)
-    steps = 2
-    step_percentage = (100/subflows)/steps
+logging.basicConfig(
+    stream=LOG_STREAM,
+    level=logging.INFO,
+    format="%(asctime)s --- %(name)s --- %(levelname)s --- %(message)s",
+    datefmt="%Y-%m-%dT%H:%M:%S%z",
+)
+
+LOGGER = logging.getLogger("script")
+
+TABLE_TITLES = {
+    "netflix_ratings": props.Translatable(
+        {
+            "en": "Ratings you gave according to Netlix:",
+            "nl": "Jouw beoordelingen volgens Netflix:",
+        }
+    ),
+}
+
+
+def process(session_id):
+    LOGGER.info("Starting the donation flow")
+    yield donate_logs(f"{session_id}-tracking")
 
     # progress in %
+    subflows = 1
+    steps = 2
+    step_percentage = (100 / subflows) / steps
     progress = 0
+    progress += step_percentage
 
-    for index, platform in enumerate(platforms):
-        meta_data = []
-        meta_data.append(("debug", f"{platform}: start"))
+    platform_name = "Netflix"
+    table_list = None
 
-        # STEP 1: select the file
-        progress += step_percentage
-        data = None
-        while True:
-            meta_data.append(("debug", f"{platform}: prompt file"))
-            promptFile = prompt_file(platform, "application/zip, text/plain")
-            fileResult = yield render_donation_page(platform, promptFile, progress)
-            if fileResult.__type__ == 'PayloadString':
-                meta_data.append(("debug", f"{platform}: extracting file"))
-                extractionResult = doSomethingWithTheFile(platform, fileResult.value)
-                if extractionResult != 'invalid':
-                    meta_data.append(("debug", f"{platform}: extraction successful, go to consent form"))
-                    data = extractionResult
-                    break
-                else:
-                    meta_data.append(("debug", f"{platform}: prompt confirmation to retry file selection"))
-                    retry_result = yield render_donation_page(platform, retry_confirmation(platform), progress)
-                    if retry_result.__type__ == 'PayloadTrue':
-                        meta_data.append(("debug", f"{platform}: skip due to invalid file"))
-                        continue
+    while True:
+        LOGGER.info("Prompt for file for %s", platform_name)
+        yield donate_logs(f"{session_id}-tracking")
+
+        promptFile = prompt_file("application/zip, text/plain", platform_name)
+        file_result = yield render_donation_page(platform_name, promptFile, progress)
+        selected_user = ""
+
+        if file_result.__type__ == "PayloadString":
+            validation = netflix.validate_zip(file_result.value)
+
+            # Flow logic
+            # Happy flow: Valid DDP, user was set selected
+            # Retry flow 1: No user was selected, cause could be for multiple reasons see code
+            # Retry flow 2: No valid Netflix DDP was found
+            # Retry flows are separated for clarity and you can provide different messages to the user
+
+            if validation.ddp_category is not None:
+                LOGGER.info("Payload for %s", platform_name)
[…]
