### 2023-09-17 ce25d7b
Added pytest for automated testing of scripts
