+++ src/framework/processing/py/pyproject.toml
@@ -8,7 +8,8 @@ authors = ["Emiel van der Veen <e.vanderveen@eyra.co>"]
 python = "^3.7"
 panda = "^0.3.1"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.test.dependencies]
+pytest = "^7.4.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
