### 2023-09-17 eff2988
Made loading of Python script use relative path
This ensures it loads when the HTML is loaded from a sub-folder.
