+++ src/framework/processing/py_worker.js
@@ -95,7 +95,7 @@ function installPortPackage() {
   console.log('[ProcessingWorker] load port package')
   return self.pyodide.runPythonAsync(`
     import micropip
-    await micropip.install("/port-0.0.0-py3-none-any.whl", deps=False)
+    await micropip.install("../../port-0.0.0-py3-none-any.whl", deps=False)
     import port
   `);  
 }
