+++ src/framework/visualisation/react/ui/workers/visualizationDataFunctions/prepareChartData.ts
@@ -6,13 +6,18 @@ import {
   ChartVisualization
 } from '../../../../../types/visualizations'
 
-export async function prepareChartData (
+export async function prepareChartData(
   table: PropsUITable & TableContext,
   visualization: ChartVisualization
 ): Promise<ChartVisualizationData> {
   const visualizationData: ChartVisualizationData = {
     type: visualization.type,
-    xKey: { label: visualization.group.label !== undefined ? visualization.group.label : visualization.group.column },
+    xKey: {
+      label:
+        visualization.group.label !== undefined
+          ? visualization.group.label
+          : visualization.group.column
+    },
     yKeys: {},
     data: []
   }
@@ -26,7 +31,9 @@ export async function prepareChartData (
   // KASPER CHECK: I think the first clause in the statement can go
   // getTableColumn will return a string array or errs out
   // so only check for length is still doing something
-  if (groupBy.length === 0) { throw new Error(`X column ${table.id}.${visualization.group.column} not found`) }
+  if (groupBy.length === 0) {
+    throw new Error(`X column ${table.id}.${visualization.group.column} not found`)
+  }
   let xSortable: Array<string | number> | null = null // separate variable allows using epoch time for sorting dates
 
   // ADD CODE TO TRANSFORM TO DATE, BUT THEN ALSO KEEP AN INDEX BASED ON THE DATE ORDER
@@ -49,15 +56,16 @@ export async function prepareChartData (
 
     // if missing values should be treated as zero, we need to add the missing values after knowing all groups
     const addZeroes = value.addZeroes ?? false
-    const groupSummary: Record<string, any> = {}
+    const groupSummary: Record<string, { n: number; sum: number }> = {}
     const uniqueGroups = new Set<string>([])
 
     for (let i = 0; i < groupBy.length; i++) {
       const xValue = groupBy[i]
       const yValue = yValues[i]
-      const group = (yGroup != null) ? yGroup[i] : (value.label !== undefined ? value.label : value.column)
+      const group =
+        yGroup != null ? yGroup[i] : value.label !== undefined ? value.label : value.column
       if (addZeroes) uniqueGroups.add(group)
-      const sortBy = (xSortable != null) ? xSortable[i] : groupBy[i]
+      const sortBy = xSortable != null ? xSortable[i] : groupBy[i]
 
       // calculate group summary statistics. This is used for the mean, pct and count_pct aggregations
       if (groupSummary[group] === undefined) groupSummary[group] = { n: 0, sum: 0 }
@@ -85,7 +93,9 @@ export async function prepareChartData (
 
       if (!aggregate[xValue][group]) aggregate[xValue][group] = 0
       if (aggFun === 'count' || aggFun === 'count_pct') aggregate[xValue][group] += 1
-      if (aggFun === 'sum' || aggFun === 'mean' || aggFun === 'pct') { aggregate[xValue][group] += Number(yValue) || 0 }
+      if (aggFun === 'sum' || aggFun === 'mean' || aggFun === 'pct') {
+        aggregate[xValue][group] += Number(yValue) || 0
+      }
     }
 
     Object.keys(groupSummary).forEach((group) => {
@@ -94,9 +104,15 @@ export async function prepareChartData (
           if (addZeroes) aggregate[xValue][group] = 0
           else continue
         }
-        if (aggFun === 'mean') { aggregate[xValue][group] = aggregate[xValue][group] / groupSummary[group].n }
-        if (aggFun === 'count_pct') { aggregate[xValue][group] = (100 * aggregate[xValue][group]) / groupSummary[group].n }
[…]
