### 2023-09-22 #5596 @osorensen
@lauraboeschoten, could you please go through this list, see if the suggestions are correct, and if so add missing DOIs?

```
- 10.1126/science.1197872 may be a valid DOI for title: Ensuring the data-rich future of the social sciences
- 10.1080/13645579.2019.1682840 may be a valid DOI for title: Digital methods in a post-API environment
- 10.4324/9781003206972-2 may be a valid DOI for title: After the ‘APIcalypse’: Social media platforms and their fight against critical scholarly research
- 10.31235/osf.io/56f4q may be a valid DOI for title: Computational research in the post-API age
- 10.5117/ccr2022.2.002.boes may be a valid DOI for title: A framework for privacy preserving digital trace data collection through data donation
- 10.31219/osf.io/krqb9 may be a valid DOI for title: Promises and pitfalls of social media data donations
- 10.3233/ds-210035 may be a valid DOI for title: Automatic de-identification of data download packages
- 10.1177/07591063211061754 may be a valid DOI for title: Which is your favorite music genre? A validity comparison of Facebook data and survey data
- 10.31235/osf.io/xjk6t may be a valid DOI for title: OSD2F: An open-source data donation framework
- 10.1016/j.patter.2022.100444 may be a valid DOI for title: Privacy-preserving local analysis of digital trace data: A proof-of-concept
```
