### 2023-09-22 #5596 @osorensen
@lauraboeschoten, I'll post a few additional comments on the paper in this thread. Here is the first one.

"Figure 1" appears twice in Figure 1. Please fix.
