### 2023-09-23 be16f3b
Added embedded mode (iframe)
