+++ dist/framework/visualisation/react/ui/elements/spinner.d.ts
@@ -1,6 +1,0 @@
-/// <reference types="react" />
-import { Weak } from '../../../../helpers';
-import { PropsUISpinner } from '../../../../types/elements';
-declare type Props = Weak<PropsUISpinner>;
-export declare const Spinner: ({ spinning, color }: Props) => JSX.Element;
-export {};
