### 2023-09-26 b4e3cac
merged with kasper, reset to old script.py files, bumped ES5 to ES6 as a compile target
