+++ dist/framework/visualisation/react/ui/pages/templates/page.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="react" />
 interface PageProps {
     body: JSX.Element;
-    sidebar: JSX.Element;
+    sidebar?: JSX.Element;
     footer: JSX.Element;
 }
 export declare const Page: (props: PageProps) => JSX.Element;
