+++ dist/framework/visualisation/react/ui/prompts/consent_form.d.ts
@@ -2,6 +2,6 @@
 import { Weak } from '../../../../helpers';
 import { PropsUIPromptConsentForm } from '../../../../types/prompts';
 import { ReactFactoryContext } from '../../factory';
-declare type Props = Weak<PropsUIPromptConsentForm> & ReactFactoryContext;
+type Props = Weak<PropsUIPromptConsentForm> & ReactFactoryContext;
 export declare const ConsentForm: (props: Props) => JSX.Element;
 export {};
