+++ dist/framework/visualisation/react/ui/elements/table.d.ts
@@ -1,10 +1,12 @@
 /// <reference types="react" />
-import { Weak } from '../../../../helpers';
-import { PropsUITable, PropsUITableRow } from '../../../../types/elements';
-import { ReactFactoryContext } from '../../factory';
-declare type Props = Weak<PropsUITable> & TableContext & ReactFactoryContext;
-export interface TableContext {
-    onChange: (id: string, rows: PropsUITableRow[]) => void;
+import { TableWithContext } from '../../../../types/elements';
+export interface Props {
+    table: TableWithContext;
+    show: boolean;
+    locale: string;
+    search: string;
+    handleDelete?: (rowIds: string[]) => void;
+    handleUndo?: () => void;
+    pageSize?: number;
 }
-export declare const Table: ({ id, head, body, readOnly, pageSize, locale, onChange }: Props) => JSX.Element;
-export {};
+export declare const Table: ({ table, show, locale, search, handleDelete, handleUndo, pageSize }: Props) => JSX.Element;
