+++ dist/framework/visualisation/react/ui/pages/donation_page.d.ts
@@ -2,6 +2,6 @@
 import { Weak } from '../../../../helpers';
 import { PropsUIPageDonation } from '../../../../types/pages';
 import { ReactFactoryContext } from '../../factory';
-declare type Props = Weak<PropsUIPageDonation> & ReactFactoryContext;
+type Props = Weak<PropsUIPageDonation> & ReactFactoryContext;
 export declare const DonationPage: (props: Props) => JSX.Element;
 export {};
