+++ dist/framework/types/pages.js
@@ -2,7 +2,8 @@ import { isInstanceOf } from '../helpers';
 export function isPropsUIPage(arg) {
     return (isPropsUIPageSplashScreen(arg) ||
         isPropsUIPageDonation(arg) ||
-        isPropsUIPageEnd(arg));
+        isPropsUIPageEnd(arg) ||
+        isPropsUIPageError(arg));
 }
 export function isPropsUIPageSplashScreen(arg) {
     return isInstanceOf(arg, 'PropsUIPageSplashScreen', []);
@@ -12,4 +13,7 @@ export function isPropsUIPageDonation(arg) {
 }
 export function isPropsUIPageEnd(arg) {
     return isInstanceOf(arg, 'PropsUIPageEnd', []);
+}
+export function isPropsUIPageError(arg) {
+    return isInstanceOf(arg, 'PropsUIPageError', ['stacktrace']);
 }

[…]
