+++ D3I_JOSS_paper/paper.md
@@ -67,7 +67,7 @@ An alternative approach to gain access to digital traces is enabled thanks to th
 
 This legislation allows researchers to invite participants to share their DDPs. A major challenge is, however, that DDPs potentially contain very sensitive data. Conversely, often not all data is needed to answer the specific research question. To tackle these challenges, @boeschoten2022framework developed an alternative workflow:  First, the participant requests their personal DDP at the platform of interest. Second, they download it onto their own personal device. Third, by means of *local processing*, only the features of interest to the researcher are extracted from that DDP. Fourth, the participant inspects the extracted features after which they can choose what they want to donate (or decline to donate). Only after selecting the data for donation and clicking the button *donate*, the donated data is sent to a storage location and can be accessed by the researcher. See \autoref{fig:workflow} for an overview of these steps.
 
-![Figure 1: An overview of the participant's data donation flow as presented by @boeschoten2022framework.\label{fig:workflow}](data_donation_flow.pdf)
+![An overview of the participant's data donation flow as presented by @boeschoten2022framework.\label{fig:workflow}](data_donation_flow.pdf)
 
 In recent years, researchers have implemented the local processing step in multiple ways. Studies by @van2022promises and @kmetty2022your used an approach where DDPs were donated directly, so without a local processing step taking place. However, extensive de-identification procedures were in place to guarantee participant’s privacy (see e.g. @boeschoten2021automatic). In addition, multiple apps have been developed to enable this local processing step. For example, the app OSD2F [@araujo2022osd2f] allows to select .json files within the DDP and performs certain de-identification steps prior to donation. Alternatively, the app Port [@boeschoten2022privacy] allows for more rigorous data processing, as researchers can write a custom Python script that locally performs any requested processing task. While this approach undoubtedly offers greater flexibility, it concurrently places a larger level of responsibility on the researcher. Other apps have been documented in academic literature. However, the feasibility of their reuse for researchers outside their respective institutions or for other platforms as intended is more ambiguous. Examples of such apps are Designerly Data Donation [@GomezOrtega_2021], the Data Donation Module [@Pfiffner_2022], the Social Media Donator [@Zannettou_2023] and WhatsR [@Kohne_2023].
 
@@ -75,9 +75,9 @@ In this paper, we introduce a new version of Port. It is open-source and allows 
 
 # Which digital platforms are investigated?
 
-Port is a tool that allows researchers to collect digital traces through donation of DDPs. In practice, this means that Port can be configured to process DDPs from any data controller, i.e, any legal entity that processes personal data. However, collection of digital traces through data donation using Port can only be a viable approach for data collection if the platform acting as data controller meets certain criteria. 
+Port is a tool that allows researchers to collect digital traces through donation of DDPs. In practice, this means that Port can be configured to process DDPs from any data controller, i.e., any legal entity that processes personal data. However, collection of digital traces through data donation using Port can only be a viable approach for data collection if the platform acting as data controller meets certain criteria. 
 
-First, in order for data donation research to occur, a platform must comply with the individual data access request that was submitted to it, meaning that platform compliance with the GDPR is a condition sine qua non for effective data donation research. Second, the process to request a copy of one's personal data should be standardized to a certain extent such that researchers can provide study participants with instructions on how to do this. Third, the file format of the DDP should ideally have a certain level of standardization as well. It is not possible to plan the procedure or extraction data from the DDP if it is unknown to the researcher where the data of interest can be found within the DDP.  
[…]
