+++ D3I_JOSS_paper/paper.bib
@@ -17,6 +17,7 @@
   number={6018},
   pages={719--721},
   year={2011},
+  doi={10.1126/science.1197872},
   publisher={American Association for the Advancement of Science}
 }
 
@@ -28,6 +29,7 @@
   number={3},
   pages={277--290},
   year={2020},
+  doi={10.1080/13645579.2019.1682840},
   publisher={Taylor \& Francis}
 }
 
@@ -39,6 +41,7 @@
   number={11},
   pages={1544--1566},
   year={2019},
+  doi={10.1080/1369118X.2019.1637447},
   publisher={Taylor \& Francis}
 }
 
@@ -50,6 +53,7 @@
   number={4},
   pages={665--668},
   year={2018},
+  doi={10.31235/osf.io/56f4q},
   publisher={Taylor \& Francis}
 }
 
@@ -61,6 +65,7 @@
   number={2},
   pages={388--423},
   year={2022},
+  doi={10.5117/ccr2022.2.002.boes},
   publisher={Amsterdam University Press Amsterdam}
 }
 
@@ -84,6 +89,7 @@
   number={4},
   pages={266--282},
   year={2022},
+  doi={10.31219/osf.io/krqb9},
   publisher={Taylor \& Francis}
 }
 
@@ -95,6 +101,7 @@
   number={2},
   pages={101--120},
   year={2021},
+  doi={10.3233/ds-210035},
   publisher={IOS Press}
 }
 
@@ -106,6 +113,7 @@
   number={1},
   pages={82--104},
   year={2022},
+  doi={10.1177/07591063211061754},
   publisher={SAGE Publications Sage UK: London, England}
 }
 
@@ -117,6 +125,7 @@
   number={2},
   pages={372--387},
   year={2022},
+  doi={10.31235/osf.io/xjk6t},
   publisher={Amsterdam University Press Amsterdam}
 }
 
@@ -128,6 +137,7 @@
   number={3},
   pages={100444},
   year={2022},
+  doi={10.1016/j.patter.2022.100444},
   publisher={Elsevier}
 }
 
