### 2023-09-28 #5596 @lauraboeschoten
Hi @osorensen , I checked the DOIs, included them in the bibfile and made the minor adjustments in the text, the latest version can be found [here](https://github.com/eyra/port/tree/v1.0.1-d3I-jboss-paper/D3I_JOSS_paper). Thank you and the reviewers for your thorough feedback!
