### 2023-10-02 #5596 @osorensen
@lauraboeschoten, do I understand correctly that the revised paper in the branch `v1.0.1-d3I-jboss-paper`? If so, I'll have to ask @openjournals/dev for some technical help before I can proceed with this submission.
