### 2023-10-02 #5596 @xuanxu
The bot fails to checkout the `v1.0.1-d3I-jboss-paper` branch because it needs a unique git reference but there is a [tag with the same name](https://github.com/eyra/port/releases/tag/v1.0.1-d3I-jboss-paper) as the branch.
The easiest ways to solve the problem would be having a separate branch for the review with a unique name or rename/delete that tag.
