### 2023-10-03 #5596 @trbKnl
@osorensen

> Make a tagged release of your software, and list the version tag of the archived version here.

tag: v1.0.1

> Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

https://zenodo.org/record/8401806

> Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

should be ✅

> Please list the DOI of the archived version here.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8401806.svg)](https://doi.org/10.5281/zenodo.8401806)
