### 2023-10-03 #5596 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
