### 2023-10-03 #5596 @oliviaguest
Huge thanks to the reviewers @kaustavbhattacharjee, @leonardojaneisand editor @osorensen! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the authors @lauraboeschoten! 🥳 🍾
